Remove redundant emptiness check from genericParseNumber https://bugs.webkit.org/show_bug.cgi?id=285184 rdar://142083712
Reviewed by NOBODY (OOPS!).

Merge: https://source.chromium.org/chromium/chromium/src/+/ab685d36561909a7ef373183f37fc3d6fc21bb83

When the 'start = buffer.position()' condition was reached we could be sure that we
had already consumed at least one character - one of '.' or '0'-'9' or
whitespace (potentially also '+'/'-' as a prefix) - and that even holds
true moving the definition of |start| after consuming any leading
whitespace. The reason for this is the up-front check for any character
in the set '0'-'9' or '.'.
Remove the redundant checks - replacing them with an assert - while
moving the definition of 'start' so that it doesn't point before any
leading whitespace.

* Source/WebCore/svg/SVGParserUtilities.cpp:
(WebCore::genericParseNumber):