Percentage height becomes auto if containing block height is auto
https://bugs.webkit.org/show_bug.cgi?id=289143
rdar://146207504

Reviewed by NOBODY (OOPS!).

This patch aligns WebKit with Blink / Chromium.

Merge: https://chromium-review.googlesource.com/c/chromium/src/+/1617833

It's wrong to walk all the way to the root and look for something with a
fixed block size just to determine whether we can collapse through the
block or not. It only matters what the actual block size of the
containing block is.

This bug triggered  because the cached state of self-collapsing got out of sync
with reality, because some ancestor went from auto-height to fixed-height, a change
that shouldn't affect the descendant at all.

* Source/WebCore/rendering/RenderBlock.cpp:
(WebCore::RenderBlock::isSelfCollapsingBlock const):
* LayoutTests/imported/w3c/web-platform-tests/css/CSS2/normal-flow/margin-collapse-through-percentage-height-block.html:
* LayoutTests/imported/w3c/web-platform-tests/css/CSS2/normal-flow/margin-collapse-through-percentage-height-block-expected.html: