[LegacySVG] The SVG image has incorrect size when it has fixed size and preserveAspectRatio="none" but the container has percent size
https://bugs.webkit.org/show_bug.cgi?id=270550
rdar://124106110

Reviewed by NOBODY (OOPS!).

This currently fixes the issue only in LegacySVG engine and LBSE has
different issue in similar area. It is good to add tests now for future tracking.

Merge: https://chromium.googlesource.com/chromium/blink/+/a7bcf4b14d5f4734039492e28ec3d84fc1d9e00a ,
https://chromium.googlesource.com/chromium/blink/+/515b1afb793f3a081f409776a818371130d9b9da ,
https://chromium.googlesource.com/chromium/src.git/+/317042f2f80c911a08a37ca27d8916f3989a6c71 ,
https://chromium.googlesource.com/chromium/src.git/+/f5404e4bd4aa7f6aaf919cb44adbaebf4b974601

When referencing an SVG image from <svg:image> with pAR set to 'none', and
said image had no intrinsic size, but a 'viewBox', the use of the intrinsic
size (which would be computed as 300x150) as the container size would
incorrect. The 'viewBox' would resolve against 300x150, so if it defined
a different aspect ratio than that, the image would not appear correctly
scaled.

For pAR=none there're a number of cases to consider:
  1) The referenced image has a 'viewBox'.
  2) The referenced image has intrinsic dimensions.
  3) The referenced image has none of the above.
For cases (1) and (2), we should use the 'viewBox' and intrinsic dimensions
(respectively) to define the container size (~= the image's viewport maps
to the viewport defined by <svg:image>). In case (3) we try to use whatever
is left to use (300x150). (In practice cases (2) and (3) should be
equivalent.) Similarly, it fixes repainting issue around resize and scaling.

Lastly, if a referenced image had the same dimensions as the referencing <image>,
a container size would not be set. This could cause the wrong scaling (etc)
to be applied if there existed another reference to the same image resource
that ended up applying a different scale-factor.
To ensure that a container size is set in this case, determine if the
referenced image has a container size set or not by checking if what
the "base" image (as returned by CachedImage::image()) differs from the
renderer-specific image (as returned by CachedImage::imageForRenderer)

* Source/WebCore/rendering/svg/legacy/LegacyRenderSVGImage.cpp:
(WebCore::containerSizeIsSetForRenderer):
(WebCore::LegacyRenderSVGImage::computeImageViewportSize const):
(WebCore::LegacyRenderSVGImage::updateImageViewport):
* Source/WebCore/rendering/svg/legacy/LegacyRenderSVGImage.h:
* LayoutTests/svg/custom/svg-image-par-resize.html:
* LayoutTests/svg/custom/svg-image-par-resize-expected.html:
* LayoutTests/svg/custom/svg-image-par-none-zero-intrinsic-size.html:
* LayoutTests/svg/custom/svg-image-par-none-zero-intrinsic-size-expected.html:
* LayoutTests/svg/custom/svg-image-par-none-no-intrinsic-size.html:
* LayoutTests/svg/custom/svg-image-par-none-no-intrinsic-size-expected.html:
* LayoutTests/svg/custom/svg-image-container-size.html:
* LayoutTests/svg/custom/svg-image-container-size-expected.html: