option element should handle empty label attribute and also should not trim label value https://bugs.webkit.org/show_bug.cgi?id=292994 rdar://151309514
Reviewed by NOBODY (OOPS!).

This patch aligns WebKit with Gecko / Firefox and Blink / Chromium.

This patch removes following cases, where we were not handling `empty`
label attribute case and were ignoring it and also trimming label value.

While web specification [1] does not state any such thing:

[1] https://html.spec.whatwg.org/multipage/form-elements.html#concept-option-label

"The label of an option element is the value of the label content attribute,
if there is one and its value is not the empty string, or, otherwise,
the value of the element's text IDL attribute."

Similarly, following issue was also raised to clarify behavior and relevant
WPT tests were added [2], which we now progress with this patch:

[2] https://github.com/whatwg/html/issues/10955

* LayoutTests/TestExpectations:
* Source/WebCore/html/HTMLOptionElement.cpp:
(WebCore::HTMLOptionElement::label const):