Parsing stop-color with invalid values (i.e., "1234" - hashless color value) https://bugs.webkit.org/show_bug.cgi?id=265420
Reviewed by NOBODY (OOPS!).

This patch aligns WebKit with Gecko / Firefox and Blink / Chromium.

Currently, while trying to parse 'stop-color', we trip in 'parseNumericColor'
function, where Quirk Mode behavior lead to inadverently applying quirky
color to 'stop-color' in SVG attribute mode as well.

This patch adds additional check to ensure that we are not in 'StrictMode',
which also cover 'SVG' attribute mode.

* LayoutTests/imported/w3c/web-platform-tests/svg/pservers/parsing/stop-color-invalid-expected.txt:
* Source/WebCore/css/parser/CSSParserFastPaths.cpp:
(WebCore::parseNumericColor):