The `initial-value` descriptor when registering Custom Properties can be optional https://bugs.webkit.org/show_bug.cgi?id=276223 rdar://131288198
Reviewed by NOBODY (OOPS!).

This patch aligns WebKit with Web Specification [1]:

[1] https://drafts.css-houdini.org/css-properties-values-api/#initial-value-descriptor

"If the value of the syntax descriptor is the universal syntax definition, then
the initial-value descriptor is optional. If omitted, the initial value of the
property is the guaranteed-invalid value."

This patch removes early return in cases of parsing declaration value and initial
value for universal syntax.

* Source/WebCore/css/parser/CSSVariableParser.cpp:
(WebCore::CSSVariableParser::parseDeclarationValue):
(WebCore::CSSVariableParser::parseInitialValueForUniversalSyntax):