Fix SVGTextContentElement lengthAdjust not resetting when attribute removed https://bugs.webkit.org/show_bug.cgi?id=304567 rdar://166967295
Reviewed by NOBODY (OOPS!).

When the lengthAdjust attribute is removed from an SVG text content element,
the baseVal property should reset to the default value (SVGLengthAdjustSpacing).
Previously, it would remain at the last set value because the code only updated
the internal value when fromString() returned a valid enum value greater than 0.

When the attribute is removed, fromString() returns SVGLengthAdjustUnknown (0),
causing the condition to fail and leaving the old value in place.

Fix by explicitly setting the value to SVGLengthAdjustSpacing when fromString()
returns an invalid value (0), ensuring proper reset to the default state.

* Source/WebCore/svg/SVGTextContentElement.cpp:
(WebCore::SVGTextContentElement::attributeChanged):
* LayoutTests/imported/w3c/web-platform-tests/svg/types/scripted/SVGAnimatedEnumeration-initial-values-expected.txt: Progression